+++ packages/react-wildcat-prefetch/test/fixtures/stubData.js
@@ -1,0 +1,17 @@
+exports.prefetchedData = {
+    "asyncData": {
+        "stub": true
+    },
+
+    "asyncArrayData": [1, 2, 3],
+
+    "firstData": {
+        "first": true
+    },
+
+    "secondData": {
+        "second": true
+    }
+};
+
+exports.fetchPromise = () => Promise.resolve(exports.prefetchedData.asyncData);